requirements.txt
folium==0.19.5
geemap==0.35.3
geopandas==1.0.1
leafmap==0.44.0
numpy==2.2.5
osmnx==2.0.3
pandas==2.2.3
Pillow==11.2.1
protobuf==6.30.2
rasterio==1.4.3
Shapely==2.1.0
streamlit==1.45.0
streamlit_folium==0.25.0
streamlit_image_coordinates==0.2.1
